refactor: move all toJSON in AbstractMetadataTimer classes to parent class
