+++ src/main/java/com/soebes/maven/extensions/artifact/DeployTimer.java
@@ -61,33 +61,4 @@ public class DeployTimer
                      NumberFormat.getNumberInstance().format( mibPerSeconds ) );
         LOGGER.info( "------------------------------------------------------------------------" );
     }
-
-    public JSONObject toJSON()
-    {
-        JSONObject jsonObject = new JSONObject();
-
-        long totalInstallationTime = 0;
-        long totalInstallationSize = 0;
-
-        for ( Entry<String, TimePlusSize> item : this.getTimerEvents().entrySet() )
-        {
-            totalInstallationTime += item.getValue().getElapsedTime();
-            totalInstallationSize += item.getValue().getSize();
-
-            JSONObject jsonItem = new JSONObject();
-            jsonItem.put("time", item.getValue().getElapsedTime());
-            jsonItem.put("size", item.getValue().getSize());
-
-            jsonObject.put(item.getKey(), jsonItem);
-        }
-
-        double mibPerSeconds = calculateMegabytesPerSeconds( totalInstallationTime, totalInstallationSize );
-
-        jsonObject.put("time", totalInstallationTime);
-        jsonObject.put("size", totalInstallationSize);
-        jsonObject.put("rate", mibPerSeconds);
-
-        return jsonObject;
-    }
-
 }
